refactored unit tests
