+++ kroxylicious-kms-provider-hashicorp-vault-test-support/src/test/java/io/kroxylicious/kms/provider/hashicorp/vault/VaultTestKmsFacadeTest.java
@@ -35,5 +35,4 @@ class VaultTestKmsFacadeTest extends AbstractTestKmsFacadeTest<Config, String, V
             assertThat(facade.getKmsServiceConfig()).isInstanceOf(Config.class);
         }
     }
-
 }
