+++ kroxylicious-kms-provider-kroxylicious-inmemory-test-support/src/main/java/io/kroxylicious/kms/provider/kroxylicious/inmemory/InMemoryTestKekManager.java
@@ -27,7 +27,7 @@ public class InMemoryTestKekManager implements TestKekManager {
 
     @Override
     public void generateKek(String alias) {
-       Objects.requireNonNull(alias);
+        Objects.requireNonNull(alias);
 
         try {
             read(alias);
